pyproject.toml
[project]
name = "lambdalib"
version = "0.1.0"
description = "Lambdalib is a collection of lib/tools for amaranth-hdl"
authors = [
    {name = "Pierre-Olivier Vauboin", email = "po@lambdaconcept.com"},
]
dependencies = [
    "luna @ git+https://github.com/greatscottgadgets/luna.git",
    "amaranth-stdio @ git+https://github.com/amaranth-lang/amaranth-stdio@8ebce9bc535f96a0f479a53f1fa000bff71aceb8",
    "amaranth-boards @ git+https://github.com/amaranth-lang/amaranth-boards.git",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
